Add Quick Entry Buttons and Context Menu
Include quick entry buttons (one for setting the current episode on S:1,
E:1 and one for using the current url as the show's url link) on the
menu for adding entries. A context menu item is also added. When
something is selected on the page there is an option to use that
selection as the name of a new entry.